feat: ヒント機能の実装 (修正)
バイナリパズルにヒント機能を追加しました。
ロジックを`BoardNormal.ts`から`BoardBinairo.ts`に移動し、
正しいボードタイプで機能するように修正しました。

この機能は以下の要素から構成されています。
- 確定可能なセルを特定するロジック (`BoardBinairo.ts`)
- どのルール（三連禁止、バランス、ユニークネス）に基づいてセルが確定されたかを説明するポップアップ
- 確定可能なセルを視覚的にハイライトする機能

これらの機能は、`HintBtn`のクリックイベントによってトリガーされます。
